Refactor Install-WinGet.ps1 to enhance error handling, improve repository trust management, and streamline the installation process for WinGet, including better handling of settings and warnings for missing paths.

+++ _scripts/pkgs/Install-WinGet.ps1
@@ -33,11 +33,9 @@ $adminWinGetConfig = @"
 function Install-WinGet {
 
 	$PsInstallScope = "$(&{ if (Test-IsSystem) { 'AllUsers' } else { 'CurrentUser' } })"
-	$wingetOffline = (Join-Path $env:DEVBOX_HOME 'Offline\WinGet')
 
 	$nugetProviderVersion = "2.8.5.201"
 	$wingetClientVersion = "1.9.2411"
-	$wingetConfigVersion = "1.8.1911"
 
 	Invoke-ScriptSection -Title "Installing WinGet Package Manager ($PsInstallScope)" -ScriptBlock {
 
@@ -48,33 +46,43 @@ function Install-WinGet {
 		}
 		catch
 		{
-			if (-not (Get-PackageProvider | Where-Object { $_.Name -eq "NuGet" -and $_.Version -gt $nugetProviderVersion })) {
-				Write-Host ">>> Installing NuGet Package Provider: $nugetProviderVersion"
-				Install-PackageProvider -Name NuGet -MinimumVersion $nugetProviderVersion -Force -Scope $PsInstallScope
+			try 
+			{
+				if (-not (Get-PackageProvider | Where-Object { $_.Name -eq "NuGet" -and $_.Version -gt $nugetProviderVersion })) {
+					Write-Host ">>> Installing NuGet Package Provider: $nugetProviderVersion"
+					Install-PackageProvider -Name NuGet -MinimumVersion $nugetProviderVersion -Force -Scope $PsInstallScope
+				}
+	
+				Write-Host ">>> Set PSGallery as Trusted Repository"
+				Set-PSRepository -Name "PSGallery" -InstallationPolicy Trusted
+				# powershell.exe -MTA -Command "Set-PSRepository -Name PSGallery -InstallationPolicy Trusted"
+
+				if (-not (Get-Module -ListAvailable -Name Microsoft.WinGet.Client | Where-Object { $_.Version -ge $wingetClientVersion })) {
+					Write-Host ">>> Installing Microsoft.Winget.Client: $wingetClientVersion"
+					Install-Module Microsoft.WinGet.Client -Scope $PsInstallScope -RequiredVersion $wingetClientVersion
+					# powershell.exe -MTA -Command "Install-Module Microsoft.WinGet.Client -Scope $PsInstallScope -RequiredVersion $wingetClientVersion"
+				}
+	
+				Write-Host ">>> Installing/Repairing WinGet Package Manager"
+				Repair-WinGetPackageManager -Verbose
+				# powershell.exe -MTA -Command "Repair-WinGetPackageManager -Verbose"
+	
 			}
-
-			Write-Host ">>> Set PSGallery as Trusted Repository"
-			Set-PSRepository -Name "PSGallery" -InstallationPolicy Trusted
-			powershell.exe -MTA -Command "Set-PSRepository -Name PSGallery -InstallationPolicy Trusted"
-
-			if (-not (Get-Module -ListAvailable -Name Microsoft.WinGet.Client | Where-Object { $_.Version -ge $wingetClientVersion })) {
-				Write-Host ">>> Installing Microsoft.Winget.Client: $wingetClientVersion"
-				Install-Module Microsoft.WinGet.Client -Scope $PsInstallScope -RequiredVersion $wingetClientVersion
-				powershell.exe -MTA -Command "Install-Module Microsoft.WinGet.Client -Scope $PsInstallScope -RequiredVersion $wingetClientVersion"
+			finally 
+			{
+				Write-Host ">>> Revert PSGallery to Untrusted Repository"
+				Set-PSRepository -Name "PSGallery" -InstallationPolicy Untrusted
+				# powershell.exe -MTA -Command "Set-PSRepository -Name PSGallery -InstallationPolicy Untrusted"
 			}
-
-			Write-Host ">>> Installing/Repairing WinGet Package Manager"
-			powershell.exe -MTA -Command "Repair-WinGetPackageManager -Verbose"
-
-			Write-Host ">>> Revert PSGallery to Untrusted Repository"
-			Set-PSRepository -Name "PSGallery" -InstallationPolicy Untrusted
-			powershell.exe -MTA -Command "Set-PSRepository -Name PSGallery -InstallationPolicy Untrusted"
 		}
+		
 
 		$winget = Get-Command -Name 'winget' -ErrorAction SilentlyContinue | Select-Object -ExpandProperty Source
 		if ($winget) {
 			$wingetVersion = Invoke-CommandLine -Command $winget -Arguments "--version" | Select-Object -ExpandProperty Output 
 			Write-Host ">>> WinGet Installed: $wingetVersion"
+		} else {
+			Write-Warning "!!! WinGet still unavailable"
 		}
 	}
 
@@ -90,14 +98,16 @@ function Install-WinGet {
 				"%LOCALAPPDATA%\Microsoft\WinGet\Settings\settings.json"
 			) 
 			
-			$paths | ForEach-Object { [System.Environment]::ExpandEnvironmentVariables($_) } | Where-Object { Test-Path (Split-Path -Path $_ -Parent) -PathType Container } | ForEach-Object { 
+			$paths | ForEach-Object { [System.Environment]::ExpandEnvironmentVariables($_) } | ForEach-Object { 
 
-				Write-Host ">>> Patching WinGet ($winGetVersion) Settings: $_"
-				$adminWinGetConfig | Out-File $_ -Encoding ASCII -Force 
-				
+				if (Test-Path (Split-Path -Path $_ -Parent) -PathType Container) {
+					Write-Host ">>> Patching WinGet Settings: $_"
+					$adminWinGetConfig | Out-File $_ -Encoding ASCII -Force 
+				} else {
+					Write-Warning "!!! WinGet Settings not found: $_"
+				}
 			}
 		} 
-
 	} 
 }
 
@@ -112,25 +122,21 @@ if ($winget) {
 	if (Test-IsPacker) {
 
 		# if executed by Packer elevate the script to SYSTEM and install WinGet for all users
-		Invoke-CommandLine -Command "powershell" -Arguments "-NoLogo -Mta -ExecutionPolicy Bypass -WindowStyle Hidden -File `"$($MyInvocation.MyCommand.Path)`"" -AsSystem `
+		Invoke-CommandLine -Command "powershell" -Arguments "-NoLogo -Mta -File `"$($MyInvocation.MyCommand.Path)`"" -AsSystem `
 			| Select-Object -ExpandProperty Output `
 			| Write-Host
 		
-		$paths = @(
-			[System.Environment]::GetEnvironmentVariable("Path","Machine"),
-			[System.Environment]::GetEnvironmentVariable("Path","User")
-		)
+		# $paths = @(
+		# 	[System.Environment]::GetEnvironmentVariable("Path","Machine"),
+		# 	[System.Environment]::GetEnvironmentVariable("Path","User")
+		# )
 
-		# update the PATH environment variable to include the WinGet installation path
-		$env:Path = $paths -join ';'
+		# # update the PATH environment variable to include the WinGet installation path
+		# $env:Path = $paths -join ';'
 
-		# get the WinGet executable path - this should be available after the PATH update
-		$winget = Get-Command -Name 'winget' -ErrorAction SilentlyContinue | Select-Object -ExpandProperty Source
+		# # get the WinGet executable path - this should be available after the PATH update
+		# $winget = Get-Command -Name 'winget' -ErrorAction SilentlyContinue | Select-Object -ExpandProperty Source
 	}
 
-	if (-not ($winget)) {
-
-		# if WinGet is not installed install it for the current user
-		Install-WinGet
-	}
+	Install-WinGet
 }
